Add more expected failures on hurd-i386
These tests were run directly before, now they are run using the
test-skeleton.c infrastructure (upstream commit 0035851c3c13f). The
tests fail when running in a thread (now the default), but succeed
when running directly (with --direct). This is a bug in the thread
library, but therefore not a regression.